+++ pages/wasm32/labyrinth-game.js
@@ -1776,51 +1776,51 @@ function getImports() {
         const ret = wasm.memory;
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1502 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1501 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1504 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1503 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1506 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1505 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_39);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1508 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1507 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1510 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1509 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1512 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1511 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1514 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1513 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1516 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1515 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper1518 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper1517 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 763, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper2265 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper2264 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 1191, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper20843 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper20842 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 14041, __wbg_adapter_57);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper48241 = function(arg0, arg1, arg2) {
+    imports.wbg.__wbindgen_closure_wrapper48240 = function(arg0, arg1, arg2) {
         const ret = makeMutClosure(arg0, arg1, 32821, __wbg_adapter_60);
         return addHeapObject(ret);
     };